pyproject.toml
[project]
name = "sai-sundhar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-postgres>=0.0.17",
    "pgvector>=0.3.6",
    "python-dotenv>=1.2.2",
]
